requirements.txt
langchain==0.3.23
langchain-openai==0.3.12
openai==1.70.0
chromadb==1.0.0
tiktoken==0.9.0
python-dotenv==1.1.0
pypdf==5.4.0 